+++ src/Faithlife.Testing.RabbitMq/AssemblyInfo.cs
@@ -1,0 +1,4 @@
+using System.Runtime.CompilerServices;
+
+[assembly: InternalsVisibleTo("Faithlife.Testing.Tests")]
+[assembly: InternalsVisibleTo("DynamicProxyGenAssembly2")]